patch_management/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
paramiko==3.4.0
pydantic==2.6.1
pydantic-settings==2.1.0
alembic==1.13.1
psutil==5.9.8
prometheus-client==0.19.0
python-json-logger==2.0.7 